feat: :sparkles: IMember, ILevel 타입 추가 (구 api 타입 제거)
