+++ src/types-old/index.ts
@@ -1,6 +1,4 @@
-export type { IMember } from './member';
 export type { IFlavor } from './flavor';
-export type { IUser, IUserLevel } from './user';
 export type { IFeel } from './feel';
 export type { IReview } from './review';
 export type { ILevel } from './level';
