--- src/types-old/member.ts
@@ -1,4 +1,0 @@
-export interface IMember {
-  id: number;
-  name: string;
-}
