+++ src/types-old/review.ts
@@ -1,12 +1,14 @@
 import { IFeel } from './feel';
 import { IFlavor } from './flavor';
-import { IMember } from './member';
 
 export interface IReview {
   id?: number;
   content: string;
   feel: IFeel;
-  member: IMember;
+  member: {
+    id: number;
+    name: string;
+  };
   createdAt: string;
   flavors: IFlavor[];
 }
