+++ src/types/levels.ts
@@ -1,0 +1,6 @@
+export interface ILevel {
+  id: number;
+  exp: number;
+  imageUrl: string;
+  tier: number;
+}
