+++ src/types/members.ts
@@ -1,0 +1,9 @@
+//   TODO: authorities 타입 확인 후 추가
+export interface IMember {
+  id: string;
+  email: string;
+  levelImage: string;
+  nickname: string;
+  profileUrl: string;
+  tier: number;
+}
